Fix: When deleting expired snapshots in janitor only fetch relevant snapshot records
