+++ sqlmesh/core/state_sync/engine_adapter.py
@@ -44,7 +44,9 @@ from sqlmesh.core.snapshot import (
     SnapshotIdLike,
     SnapshotInfoLike,
     SnapshotIntervals,
+    SnapshotNameVersion,
     SnapshotNameVersionLike,
+    SnapshotTableCleanupTask,
     fingerprint_from_node,
 )
 from sqlmesh.core.snapshot.definition import (
@@ -115,6 +117,7 @@ class EngineAdapterStateSync(CommonStateSyncMixin, StateSync):
             "version": exp.DataType.build("text"),
             "snapshot": exp.DataType.build("text"),
             "kind_name": exp.DataType.build("text"),
+            "expiration_ts": exp.DataType.build("bigint"),
         }
 
         self._environment_columns_to_types = {
@@ -256,6 +259,66 @@ class EngineAdapterStateSync(CommonStateSyncMixin, StateSync):
             where=filter_expr,
         )
 
+    @transactional()
+    def delete_expired_snapshots(self) -> t.List[SnapshotTableCleanupTask]:
+        current_ts = now_timestamp(minute_floor=False)
+
+        expired_query = (
+            exp.select("name", "identifier", "version")
+            .from_(self.snapshots_table)
+            .where(exp.column("expiration_ts") <= current_ts)
+        )
+        expired_candidates = {
+            SnapshotId(name=name, identifier=identifier): SnapshotNameVersion(
+                name=name, version=version
+            )
+            for name, identifier, version in self.engine_adapter.fetchall(expired_query)
+        }
+
+        snapshots = self._get_snapshots_with_same_version(set(expired_candidates.values()))
+
+        snapshots_by_version = defaultdict(set)
+        snapshots_by_temp_version = defaultdict(set)
+        for s in snapshots:
+            snapshots_by_version[(s.name, s.version)].add(s.snapshot_id)
+            snapshots_by_temp_version[(s.name, s.temp_version_get_or_generate())].add(s.snapshot_id)
+
+        promoted_snapshot_ids = {
+            snapshot.snapshot_id
+            for environment in self.get_environments()
+            for snapshot in environment.snapshots
+        }
+
+        def _is_snapshot_used(snapshot: Snapshot) -> bool:
+            return (
+                snapshot.snapshot_id in promoted_snapshot_ids
+                or snapshot.snapshot_id not in expired_candidates
+            )
+
+        expired_snapshots = [s for s in snapshots if not _is_snapshot_used(s)]
+
+        if expired_snapshots:
+            self.delete_snapshots(expired_snapshots)
+
+        cleanup_targets = []
+        for snapshot in expired_snapshots:
+            shared_version_snapshots = snapshots_by_version[(snapshot.name, snapshot.version)]
+            shared_version_snapshots.discard(snapshot.snapshot_id)
+
+            shared_temp_version_snapshots = snapshots_by_temp_version[
+                (snapshot.name, snapshot.temp_version_get_or_generate())
+            ]
+            shared_temp_version_snapshots.discard(snapshot.snapshot_id)
+
+            if not shared_temp_version_snapshots:
+                cleanup_targets.append(
+                    SnapshotTableCleanupTask(
+                        snapshot=snapshot.table_info, dev_table_only=bool(shared_version_snapshots)
+                    )
+                )
+
+        return cleanup_targets
+
     def delete_expired_environments(self) -> t.List[Environment]:
         now_ts = now_timestamp()
         filter_expr = exp.LTE(
@@ -339,7 +402,7 @@ class EngineAdapterStateSync(CommonStateSyncMixin, StateSync):
         for where in self._snapshot_id_filter([snapshot.snapshot_id]):
             self.engine_adapter.update_table(
                 self.snapshots_table,
-                {"snapshot": _snapshot_to_json(snapshot)},
+                {"snapshot": _snapshot_to_json(snapshot), "expiration_ts": snapshot.expiration_ts},
                 where=where,
                 contains_json=True,
             )
@@ -1154,6 +1217,7 @@ def _snapshots_to_df(snapshots: t.Iterable[Snapshot]) -> pd.DataFrame:
                 "version": snapshot.version,
                 "snapshot": _snapshot_to_json(snapshot),
                 "kind_name": snapshot.model_kind_name.value if snapshot.model_kind_name else None,
+                "expiration_ts": snapshot.expiration_ts,
             }
             for snapshot in snapshots
         ]
